requirements.txt
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
kornia>=0.7.0,<1.0.0
matplotlib>=3.7.0
scikit-image>=0.21.0
tqdm>=4.65.0
pyyaml>=6.0
lpips>=0.1.4
streamlit>=1.28.0,<2.0.0
